pyproject.toml
[project]
name = "smart-contract-data"
version = "0.1.0"
description = "Data collection + experiment planning for smart contract vulnerability detection model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    
    # GitHub API
    "PyGithub>=2.1.0",
    
    # Dataset platforms
    "kaggle>=1.5.16",
    "huggingface-hub>=0.19.0",
    "datasets>=2.15.0",
    
    # Data processing
    "pandas>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    
    # PDF processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    
    # Markdown processing
    "markdown>=3.5.0",
    "markdown-it-py>=3.0.0",
    
    # Rate limiting and retries
    "ratelimit>=2.2.1",
    "tenacity>=8.2.0",
    
    # Async support
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    
    # CLI
    "click>=8.1.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    
    # Scheduling
    "schedule>=1.2.0",
    "apscheduler>=3.10.0",
    
    # Logging
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crawlers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
